+++ packages/docs/liveEditorCode/trustElements/BirlesikOdemeRegulated.code.js
@@ -1,0 +1,9 @@
+() => {
+    return (
+        <>
+           <BirlesikOdemeRegulated
+            title={'Transactions processed by Birlesik Ödeme Hizmetleri'}
+            />
+        </>
+    )
+};