+++ packages/docs/liveEditorCode/trustElements/MASRegulated.code.js
@@ -1,0 +1,11 @@
+() => {
+    return (
+        <>
+           <MASRegulated
+            title={'Regulated by the Monetary Authority of Singapore (MAS)'}
+            linkText={'Learn More'}
+            href={'https://transferwise.com/help/articles/2932693/how-is-transferwise-regulated-in-each-country'}
+            />
+        </>
+    )
+};