+++ packages/docs/liveEditorCode/trustElements/TUVApproved.code.js
@@ -1,0 +1,11 @@
+() => {
+  return (
+    <>
+      <TUVApproved
+        title="TÜV approved"
+        linkText="The report"
+        href="https://transferwise.com/gb/blog/transferwise-tuv-audit-2019"
+      />
+    </>
+  )
+};