+++ packages/marketing-components/src/index.js
@@ -1,5 +1,25 @@
 export { default as VideoModal } from './videomodal';
 export { default as YouTubeVideoModal } from './youtubevideomodal';
-export { FCARegulated, Trustpilot, CanstarAward } from './trustelements';
+export {
+  ASICRegulated,
+  BirlesikOdemeRegulated,
+  BNMApproved,
+  BrazilianCorrespondent,
+  CAndEDRegulated,
+  CanstarAward,
+  Customers,
+  DIARegulated,
+  FCARegulated,
+  FINTRACRegulated,
+  FSRAApproved,
+  JPFSARegulated,
+  MASRegulated,
+  Mitsui,
+  Safe,
+  StraitsTimes,
+  Trustpilot,
+  TrustpilotUS,
+  TUVApproved,
+} from './trustelements';
 export { default as AppStoreBadge } from './appstorebadge';
 export { default as GooglePlayStoreBadge } from './googleplaystorebadge';