+++ packages/marketing-components/src/trustelements/Safe/Safe.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import Types from 'prop-types';
+
+import TrustElement from '../TrustElement';
+import SecurityIllustration from '../SecurityIllustration';
+
+const Safe = ({ title, linkText, href }) => (
+  <TrustElement
+    src={<SecurityIllustration />}
+    title={title}
+    linkText={linkText}
+    href={href}
+    shouldAnimate
+  />
+);
+
+Safe.propTypes = {
+  title: Types.string.isRequired,
+  linkText: Types.string.isRequired,
+  href: Types.string.isRequired,
+};
+
+export default Safe;