+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/transferwise/marketing-components/compare/@transferwise/marketing-docs@1.7.0...@transferwise/marketing-docs@1.8.0) (2020-09-29)
+
+
+### Features
+
+* Added trust elements using SecurityIllustration ([#22](https://github.com/transferwise/marketing-components/issues/22)) ([6aefb21](https://github.com/transferwise/marketing-components/commit/6aefb21abec2a5587ef64e6a1949796f371aece2))
+
+
+
+
+
+# [1.7.0](https://github.com/transferwise/marketing-components/compare/@transferwise/marketing-docs@1.6.0...@transferwise/marketing-docs@1.7.0) (2020-09-29)
+
+
+### Features
+
+* Adding FSRAApproved and TrustpilotUS award element ([#18](https://github.com/transferwise/marketing-components/issues/18)) ([b708f95](https://github.com/transferwise/marketing-components/commit/b708f950b44246dbc3f305d6e7e2d4e84ee9a3e4))
+
+
+
+
+
+# [1.6.0](https://github.com/transferwise/marketing-components/compare/@transferwise/marketing-docs@1.5.0...@transferwise/marketing-docs@1.6.0) (2020-09-29)
+
+
+### Features
+
+* Added various trust elements ([#13](https://github.com/transferwise/marketing-components/issues/13)) ([71907f3](https://github.com/transferwise/marketing-components/commit/71907f371beee7ecf60eef2ef60cc488986fee29))
+
+
+
+
+
+# [1.5.0](https://github.com/transferwise/marketing-components/compare/@transferwise/marketing-docs@1.4.0...@transferwise/marketing-docs@1.5.0) (2020-09-29)
+
+
+### Features
+
+* ASIC Regulated trust element ([#17](https://github.com/transferwise/marketing-components/issues/17)) ([cbf7d55](https://github.com/transferwise/marketing-components/commit/cbf7d55f1dc5f84d83a83177e25337641f037746))
+
+
+
+
+
 # [1.4.0](https://github.com/transferwise/marketing-components/compare/@transferwise/marketing-docs@1.3.0...@transferwise/marketing-docs@1.4.0) (2020-09-29)